+++ Payroll Manager/Source Files/Classes/WorkWeek.cs
@@ -4,7 +4,7 @@ namespace PayrollManager
 {
 class WorkWeek
 {
-    public WorkWeek(DayType startingDay, params double[] hours)
+    public WorkWeek(DayType startingDay, params decimal[] hours)
     {
         Days = new Workday[7];
 
@@ -17,36 +17,36 @@ class WorkWeek
             Days[i] = new Workday((DayType)weekdayNumber, hours[i]); 
         }
 
-        double week_regular             = 0;
-        double week_overtime_under_40   = 0;
-        double week_overtime_over_40    = 0;
+        decimal week_regular             = 0.00m;
+        decimal week_overtime_under_40   = 0.00m;
+        decimal week_overtime_over_40    = 0.00m;
 
-        float interval = 0.25f;
+        decimal interval = 0.25m;
         foreach (var day in Days)
         {
-            double hours_passed_in_day      = 0;
-            double regular_in_day           = 0;
-            double day_overtime_under_40    = 0;
-            double day_overtime_over_40     = 0;
-            for (float i = 0f; i < 24f; i += interval)
+            decimal hours_passed_in_day      = 0.00m;
+            decimal regular_in_day           = 0.00m;
+            decimal day_overtime_under_40    = 0.00m;
+            decimal day_overtime_over_40     = 0.00m;
+            for (decimal i = 0.00m; i < 24.00m; i += interval)
             {
                 if (hours_passed_in_day < day.TotalHours)
                 {
-                    if (week_regular < 40)
+                    if (week_regular < 40.00m)
                     {
-                        if ( hours_passed_in_day < 8)
+                        if ( hours_passed_in_day < 8.00m)
                         {
                             regular_in_day  += interval;
                             week_regular    += interval;
                         }
-                        else if ( hours_passed_in_day >= 8)
+                        else if ( hours_passed_in_day >= 8.00m)
                         {
                             day_overtime_under_40   += interval;
                             week_overtime_under_40  += interval;
                         }
                         hours_passed_in_day += interval;
                     }
-                    else if (week_regular >= 40)
+                    else if (week_regular >= 40.00m)
                     {
                         hours_passed_in_day     += interval;
                         day_overtime_over_40    += interval;
@@ -59,22 +59,22 @@ class WorkWeek
         }
         OvertimeHours += week_overtime_under_40;
         OvertimeHours += week_overtime_over_40;
-        if (week_regular <= 40)
+        if (week_regular <= 40.00m)
         {
             RegularHours = week_regular;
         }
-        else if (week_regular > 40)
+        else if (week_regular > 40.00m)
         {
-            RegularHours    = 40;
-            OvertimeHours   += week_regular - 40;
+            RegularHours    = 40.00m;
+            OvertimeHours   += week_regular - 40.00m;
         }
     }
 
     public Workday[] Days { get; }
 
-    public double RegularHours { get; }
+    public decimal RegularHours { get; }
     
-    public double OvertimeHours { get; }
+    public decimal OvertimeHours { get; }
 
     public override string ToString()
     {
